(maint) Allow Bolt::PAL.new to take boltlib_path
This commit updates Bolt::PAL's init method to allow overriding the
boltlib_path that is set as the first entry in the modulepath when executing
bolt.

Bolt requires the boltlib_path to be set to at least Bolt::PAL::BOLTLIB_PATH
in order for bolt to function properly. But in some cases (like our PE
projects that build on top of bolt) other paths need to be added alongside the
standard BOLTLIB_PATH.