(GH-2134) Add 'bolt module add' command
This adds a new `bolt module add` command that adds a new module to the
project. The command accepts a single module name, resolves its
dependencies, updates the project configuration file and `Puppetfile`,
and installs the project's modules.

If the project configuration file already includes a declaration for the
new module, Bolt will display a helpful message and exit successfully
without doing anything.

When resolving modules, Bolt will first attempt to resolve the new
module with each of the installed modules pinned to a specific version.
This is intended to prevent any installed modules from being updated
unintentionally. If resolving with this approach fails, Bolt falls back
to re-resolving all of the module declarations with the new module,
which may result in insalled modules being updated to different
versions.

Lastly, this includes a new `ModuleInstaller` class which moves most of
the logic for installing modules from the `CLI` class.

This feature is shipping dark and does not have an associated release
note.

!no-release-note