(GH-2156) Don't leak scope between targets
When compiling catalogs, we use a single scope object to capture the
current plan scope at the time of the compilation, and then merge in
target-specific information for each compile. However, the compile
method was *modifying* the single scope object to add its
target-specific info. This caused a race condition whereby the scope
could be modified by another thread and the wrong target inserted and
used for compilation.

This appears to be relatively uncommon at low levels of concurrency,
because the critical section only lasts until the scope has been
serialized to JSON to be passed to the compilation process. Once that's
happened, any external modifications won't affect the current compile.
Due to the GIL, it's impossible for multiple threads to be running that
code at the same time, so the race condition requires a context switch
to happen in the middle of execution. With high concurrency or many
targets being compiled, this is not unlikely.

We now _merge_ the target-specific data rather than inserting it
directly, so each target gets a copy of the scope. We also `freeze` the
scope for good measure, to avoid any such issues popping up in the
future. This does not have a significant impact on memory usage as each
scope hash still refers to a single copy of the actual underlying
_data_.

!bug

  * **Prevent data for the wrong target being used when compiling `apply()` blocks**
     ([#2156](https://github.com/puppetlabs/bolt/issues/2156))

    Previously, a race condition allowed `apply()` blocks to use data
    for the wrong target during compilation. This could cause targets to
    apply incorrect or invalid catalogs.