(PUP-10653) Remove dependency on win32/dir
For Puppet 7 we are dropping the `win32/dir` dependency as we only used
constants from it, which we replaced with environment variables (see:
https://github.com/puppetlabs/puppet/pull/8314).

This would become breaking in bolt when it switches to Puppet 7 and the
win32 gems are no longer pulled.

For AIO bolt, the Puppet change won't break anything since bolt uses its
own runtime which still packages the win32 gems, but it would still be
an improvement in speed if the dependency is dropped.

The `win32/dir` gem also monkey patches some `Dir` methods, most notably
`glob` and `pwd` to use backslashes which end up causing more trouble as
Ruby uses `/` as the default separator for Windows.

Switch to using the `ALLUSERSPROFILE` environment variable instead of
the `Dir::COMMON_APPDATA` constant, and remove a `Dir.pwd` workaround.

!removal

  * **Remove dependency on win32/dir**
    ([PUP-10653](https://tickets.puppetlabs.com/browse/PUP-10653))

    Replace the usage of win32/dir constants with environment variables.