(GH-2131) Add bolt module subcommands
This adds the `bolt module generate-types` and `bolt module show`
subcommands, which are analagous to `bolt puppetfile generate-types` and
`bolt puppetfile show-modules` respectively. These commands have the
exact same behavior as the puppetfile subcommands. The environment
variable `BOLT_MODULE_FEATURE` must be set to use the `bolt module`
subcommand. The puppetfile commands aren't deprecated yet - we'll do
that closer to 3.0. This also doesn't include a release note, as we're
launching the module feature dark until the entire workflow is complete.

Closes #2131

!no-release-note