(maint) Add Litmus smoke tests
This adds a new Github workflow that verifies that Litmus executes using
the latest version of Bolt. Since Litmus relies on Bolt to work, this
verifies that any changes in Bolt won't break Litmus before we merge.
This sets up a matrix of Ubuntu and Windows platforms and runs the tests
on each, where the test on Ubuntu uses a remote target and the Windows
test verifies Litmus against localhost. Both tests simply run the Litmus
acceptance tests using the Bolt version that triggered the action to run
(so the PR branch).

!no-release-note