(GH-2148) Fix JSON ordering in pester test
The ConvertTo-JSON cmdlet respects JSON ordering in Powershell 5.1 and
7.1, but not version 7.0. This caused the test that verifies passing a
JSON object to Invoke-BoltTask to fail, as the ordering of JSON
parameters was wrong and the test simply verifies the resulting Bolt
command as a string. This updates the test to accept the JSON object
elements in any order. Because users and their tasks don't care about
the order, it doesn't make sense to fix this in the code itself - just
to update the test to also not care about the order.

Closes #2148

!no-release-note