(GH-2168) Add version requirements for module declarations
This adds support for specifying version requirements for a module
declaration under the `modules` key of a project config file. To specify
a version, the module declaration should have a `version_requirement`
key. Version requirements should be valid according to semantic-puppet,
and includes support for specific versions (`1.0.0`), version shorthand
(`1.x`), and version ranges (`>= 1.0.0 < 3.0.0`).

!no-release-note