(GH-2125) Read command from a file or stdin
This adds support for reading a command from a file or `stdin` using
the `bolt command run` command. To read a command from a file,
pass the path to the file with the `@` symbol:

```shell
$ bolt command run @script.sh --targets targets
```

The filepath is expanded relative to the current working directory.

To read a command from `stdin`, pass a single dash `-` as the command
and then pipe the command to Bolt:

```shell
$ echo whoami | bolt command run - --targets targets
```

!feature

* **Read command from a file or `stdin` using `bolt command run`**
  ([#2125](https://github.com/puppetlabs/bolt/issues/2125))

  The `bolt command run` command can now read a command from a file or
  `stdin`.