(maint) Properly detect gem install again
This detection was broken since we added the secure_env_vars module as a
built-in, but didn't add it to the list of "expected" built-ins. That
caused Bolt to always believe there was "non-gem install" content
present and so it would never warn that the user had an incomplete
install.

!bug

  * **Fix warning when running from a gem install**

    Bolt again properly detects when it's being run from a gem install
    and emits a warning.