(GH-2182) Use new modulepath when 'modules' is configured
This updates `Bolt::Project` and `Bolt::Config` to use a new modulepath
when `modules` is configured in `bolt-project.yaml`. This work is part
of the new module management workflow.

When `modules` is configured, the default modulepath is `['modules']`
and the full modulepath will have `.modules` appended to it, whether
Bolt uses the default modulepath or the user-configured modulepath. If
the user-configured modulepath includes `.modules`, Bolt will error and
display a helpful message that the modulepath cannot be configured with
`.modules` as it is automatically appended.

If `modules` is not configured, Bolt's behavior remains the same and the
default modulepath is `['modules', 'site-modules', 'site']`.

This also adds new `BoltSpec::Project` library functions that will
create a temporary project. This is useful for spec tests and prevents
Bolt from unintentionally loading configuration on the system running
the tests.

This change is shipping dark until the full module management workflow
is complete.

!no-release-note