(GH-2182) Deprecate 'puppetfile' command when 'modules' is configured
This change raises an error when running `bolt puppetfile *` if
`modules` is configured in `bolt-project.yaml`. The error includes a
helpful message telling the user the command is deprecated and to use
`bolt module *` instead.

This is part of the module management workflow, which is shipping dark
until it is complete.

!no-release-note