(GH-2182) Disable 'module' command when 'modules' is not configured
This change disables the `module` command when `modules` is not
configured in `bolt-project.yaml`. When `modules` is not configured, and
the `module` command is used, Bolt will raise an error with a helpful
message telling the user to update their project configuration to manage
module dependencies.

This also updates the logic used to display help text. Previously, if
a user ran `bolt help module` or `bolt module --help`, the `module`
command help text would be displayed. This now ensures that help text is
only shown for enabled subcommands.

The module management workflow is shipping dark until the entire
workflow is completed.

!no-release-note