(GH-2188) More reliably initialize log levels
When initializing the Logging library, we previously only checked
whether it had already been initialized and skipped doing it if it had.
However, when using Bolt as a library, it's possible that other code
(for instance, the winrm gem) might set up Logging before Bolt has a
chance to, probably initializing it with the _default_ set of log
levels. That causes Bolt to fail when trying to use the `trace` level,
since no method will exist for it on the Logger objects.

We now explicitly check not whether Logging has been initialzied at all,
but whether the levels we need exist. This does allow Logging to be
initialized with _additional_ levels beyond what Bolt needs, but ensures
that Bolt's levels exist at a minimum.

This also slightly refactors how we set up some instance variables on
the Bolt::Logger module itself, to avoid having to set them in the
initialize_logging method.

!no-release-note