(GH-2127) Parallelize plan steps across targets
This implements a new (to-be-named) `parallelize` function which
accepts an array and a block and maps the given block to the array in
parallel. That is to say, the block is executed in sequence for each
element in the array, with the array elements executing in parallel.
This is in contrast to typical plan steps which execute each step in
parallel across targets and wait for all targets to finish before
continuing to the next step.

The block will wait for the sequence to finish on all elements of the
array before returning an array of Results - one per element in the
array. The results can be either the value of a `return` statement, the
result of the last line in the block (can be nil), or an error that was
raised during the block for that element. The results are in the same
order the original array was in, regardless of which elements finish the
block 'first'. Unless `catch_errors()` wraps the block (either inside or
outside), if any element errored during the block the plan will fail and
all results (successful and error) will be included in the failure
details. If `catch_errors()` wraps the block then the array will contain
Error objects for errored results, as usual.

Parallel execution is managed by the executor. When a parallel block
begins, we make a copy of the existing variable scope for each element
in the array so that any variables defined in the scope won't get
overridden by parallel execution. Then we create a Ruby Fiber for each
element of the array executing the block, and send the array of Fibers
to the executor. The executor goes into "parallel mode", where it will
resume each fiber in turn until the block has completed for all elements
of the array. Certain plan function can then be "Fiber aware" by
starting a new thread and yielding their Fiber once the thread has
started, waiting for the thread to finish before returning the result
back to the executor. This allows certain functions to not block on
other array elements continuing to execute the block, while other
functions (or just certain parts of functions, like loading tasks) will
block.

!feature

* **Experimental: Add `parallelize()` plan function**

  This adds a new function `parallelize()` that accepts an array and a
  block and executes the block on each element of the array in parallel.
  This allows plan authors to continue executing a block on other targets
  if a particular plan function may take a long time on one target. This
  features is experimental.