(GH-2184) Refactor ProjectMigrate class
This refactors the `ProjectMigrate` class so that the logic for each
step of the migration process is contained in a separate class.
The `ProjectMigrate` class has been renamed to `ProjectMigrator`.
Migration for inventory files is now handled by the
`Bolt::ProjectMigrator::Inventory` class, while migration for
configuration files is now handled by the `Bolt::ProjectMigrator::Config`
class.

The intent behind this refactor is to break up the `ProjectMigrate`
class into more discrete components, and have the class instead defer
the work to specialized migrators. This will also allow for adding a
migrator for project modules without making the `ProjectMigrate` class
even more monolithic.

!no-release-note