(GH-2184) Update 'project migrate' to migrate modules
This updates the `bolt project migrate` command to migrate a project's
modules to the new module management workflow. If the project already
configures `modules` or configures `modulepath` to a non-default
modulepath, then the migrator will do nothing.

If the project has a `Puppetfile`, the migrator will parse the file for
modules, and then prompt the user to select direct dependencies for the
project. Any modules selected as dependencies will be used to generate a
new `Puppetfile` that includes the modules and their dependencies. Bolt
will then install the `Puppetfile` to the managed moduledir `.modules`
and add the selected modules to the project configuration. Any modules
on the modulepath that are not selected as dependencies will be
consolidated into the `modules` directory, while all remaining
moduledirs will be deleted.

If the project does not have a `Puppetfile`, then Bolt will consolidate
all modules on the modulepath into the `modules` directory and delete
all remaining moduledirs.

This feature is feature flagged and is shipping dark until the module
management workflow is complete.

!no-release-note