(GH-2135) Add 'module' subcommands to pwsh cmdlets module
This adds the `bolt module` subcommands to the Powershell cmdlets module
as `Add-BoltModule`, `Register-BoltModuleTypes`, `Install-BoltModule`,
and `Get-BoltModule`. This call directly into the respective `bolt
module` subcommands and should exhibit the same behavior.

This additionally fixes up the parameters to some Bolt commands in the
powershell module, ensuring that parameters are only included if they
are valid and marking them as optional or required appropriately.

Closes #2135

!no-release-note