(maint) Have modules key accept strings
This adds support for the 'modules' key accepting strings in addition to
hashes as part of the modules array. If a module in the array is a
barestring it will be converted to the hash `{ name: <string> }`,
similar to how targets accept a string and treat it as the URI. This
provides a nice shorthand for users listing modules manually. The `bolt
module add` command will still add the hash version.

!no-release-note