(GH-2217) Explicitly pass kwargs to methods that accept them
In some places, we were passing keyword arguments as a hash as the last
argument to the method. This syntax is deprecated in Ruby 2.7 in favor
of explicitly passing them as keyword arguments using the ** operator.
Unfortunately, it's hard to programmatically find places where this
issue occurs since it depends on the definitin of the method, so this
commit is just a first stab.

!no-release-note