(GH-2226) Do not set module version to latest when migrating
This prevents a module's version requirement in `bolt-project.yaml` from
being set to the string `:latest` when migrating a project. Previously,
if a Puppetfile had a module specification with either a version
`:latest` or no version at all, the module's version requirement would
be set to `:latest`, which is not valid. Now, in these instances the
module will not have a version requirement in the `bolt-project.yaml`.

!no-release-note