(maint) Print unsuccessful migration in red
This modifies the 'Project could not be migrated completely' message
when project migration fails to be printed in red, letting the user know
at a glance that something went wrong. After migrating a project it's
easy to not read the output if it all looks successful at a glance.
Printing in red emphasizes that something in the migration needs the
users attention, and potentially manual migration.

!no-release-note