(maint) Surface Puppetfile Resolver validation errors
We currently hardcode an error message when Puppetfile Resolver
successfully resolves a Puppetfile but with validation errors. This
updates the message to print the list of validation errors when
Puppetfile Resolver encounters them.

This additionally adds a format for migrate steps that fail, coloring
them red to indicate failure but maintaining the indent.

!no-release-note