(GH-2197) Error gracefully when yaml plans fail to parse
Previously when raising an error from a plan run in Bolt we would only
check if the error had a `file` method and use that to determine whether
the error was from Puppet or not. If the error was from Puppet we could
assume other methods on the error, like line and pos. However,
when yaml plans were syntactically invalid and failed to parse this
would raise a Psych::SyntaxError, which has a `file` method but not a
`pos` method. This would cause a stacktrace when we assumed the error
was a Puppet error and tried to call `pos` on it.

This change modifies the details hash to only add each key if the
corresponding method exists on the error. We can't be sure this will
catch all relevant details from an error (for example if the `pos`
exists on an error but is accessed by a method `position`), but it will
at least not stacktrace when a non-Puppet error with any of those
methods is raised.

Closes #2197

!bug

* **Invalid YAML plans now fail gracefully** ([2197](https://github.com/puppetlabs/bolt/issue/2197))

  Previously, if a YAML plan had a syntax error Bolt would stacktrace
  due to an assumption about what methods the resulting error had. It now
  fails gracefully with the line of the error.