(GH-2234) Add moduledir to generated Puppetfile
When generating the Puppetfile from the list of modules in
`bolt-project.yaml`, we now add the line `moduledir '.modules'` to
prevent R10K and old versions of Bolt from deleting the `modules/`
directory and installing modules from the Puppetfile to the wrong
location. This essentially ensures that using the new module management
workflow is safe with legacy tools, as the Puppetfile knows to install
modules to the new modules directory `.modules/` instead of installing
to `modules/` which should now be treated as a place for users' custom
modules. This line is only added when generating a new Puppetfile using
the new module management workflow - behavior is not modified if using
`bolt project init --modules`, which still uses the old workflow.

This additionally documents configuring the modulepath in PE
environments to pick up the new modulepath.

Closes #2234

!no-release-note