(GH-2187) Add Git module support for module management feature
This adds support for Git modules to the module management feature. Git
modules can now be specified in `bolt-project.yaml` using the following
format:

```
modules:
  - git: <GIT REPO URL>
    ref: <GIT REF>
```

When a Git module is specified, Bolt will first resolve its name from
the module's `metadata.json` and resolve the SHA for the specified
`ref`. If the URL to the repo does not point to a valid module, Bolt
will error.

After resolving the Git module's name and SHA, Bolt will resolve modules
using the `puppetfile-resolver` library. If resolution succeeds, a
Puppetfile with the Git module will be generated.

This also includes a significant refactoring of the `Bolt::Puppetfile`
class and the introduction of new `Spec` and `Module` classes. These
latter classes are used to represent module specs in `bolt-project.yaml`
and resolved modules in the Puppetfile, respectively. The
`Spec` classes are now passed to the `Resolver` class, which resolves
modules from the specs and returns a `Puppetfile` object. This
`Puppetfile` object includes fully-resolved `Module` objects, which are
used to generate a Puppetfile.

> **Note:** Only modules from GitHub are supported, as the
> `puppetfile-resolver` library only supports GitHub modules.

!feature

* **Git module support for module management feature**
  ([#2187](https://github.com/puppetlabs/bolt/issues/2187))

  Git modules can now be specified in `bolt-project.yaml` and used with
  the module management feature. Only GitHub modules are supported.