Extract config parsing out of PAL
This commit updates the Bolt::PAL class to take the whole Bolt::Config object
as a single parameter rather than taking options from config as seperate
parameters.

As part of the change, the functionality of munging the modulepath to include
BOLTLIB and built-in content has moved to a read-only function in the Config
class.

All of the inputs to PAL came from config already, and this distinction -that
PAL just takes config as one object- helps define how tools that use bolt as a
library (like PE) should use and extend bolt for their needs. Specifically
this change helps other tools use Bolt without needing to redefine PAL, but
instead those tools can redefine or update the config class for their needs
and have the config sent to PAL.