(GH-2167) Add note that pry-remote and pry-nave do not work together
This adds a note to the 'Debugging Tasks' section of the docs under the
ruby section that the library we recommend, 'pry-remote', does not work
with common pry libraries 'pry-nav' and 'pry-byebug'. This issues is
[known](https://github.com/Mon-Ouie/pry-remote/issues/35#issuecomment-20290739)
and does not seem likely to be resolved.

I opted not to include this in our known issues page since this isn't an
issue with Bolt itself.

Closes #2167

!no-release-note