(GH-2207) Log when default inventory file cannot be loaded
This change updates how Bolt logs which inventory file it loaded.
Previously, if Bolt loaded the default inventory file, it would log that
it loaded the file, even if the file does not exist. Now, if the default
inventory file does not exist, Bolt will instead notify the user that it
tried to load the file, but that it does not exist.

!bug

* **Log when default inventory file cannot be loaded**
  ([#2207](https://github.com/puppetlabs/bolt/issues/2207))

  Bolt now logs that it tried but failed to load the default inventory
  file when the default inventory file does not exist. Previously, Bolt
  would log that it loaded the default inventory file, even when it was
  unable to do so.