(GH-2205) Improve 'bolt inventory show' output
This improves the output for the `bolt inventory show` command.

When using the `human` outputter, Bolt will now display each target's source
alongside the target. If the target was loaded from inventory, it will
show the path to the inventory file; if the target is an adhoc target,
it will say it was not loaded from inventory. The output will also show
where inventory was loaded from, and tell the user if the inventory file
it attempted to load does not exist. Lastly, the target count now lists
the total number of targets, the number loaded from inventory, and the
number of adhoc targets.

When using the `json` outputter, Bolt will now include `inventory` and
`adhoc` keys. The `inventory` key includes a list of targets loaded from
inventory, the number of targets loaded from inventory, and the path to
the inventory file. The `adhoc` key lists the names of adhoc targets,
and the total number of adhoc targets.

!feature

* **Improved output for `bolt inventory show` and `Get-BoltInventory`**
  ([#2205](https://github.com/puppetlabs/bolt/issues/2205))

  The `bolt inventory show` command and `Get-BoltInventory` cmdlet now
  show where each target was loaded from alongside the target's name.
  Output also includes the path to the loaded inventory file and the
  number of inventory targets and adhoc targets.