Refactor PAL to use an object to manage modulepath
This commit updates the Bolt::PAL class to use a newly created
Bolt::Config::Modulepath object to manage the modulepath that's used during
execution of PAL and the Applicator class.

The Modulepath object will be responsibile for holding the value of the
modulepath and providing any transformations (like adding BOLTLIB_PATH) for
use with other parts of bolt.

!no-release-note