(GH-2230) Print Puppetfile diff when using 'module add'
This adds a new message to the outputter that describes changes made to
a Puppetfile when using the `bolt module add` / `Add-BoltModule`
command. The message will display which modules have been added,
removed, upgraded, and downgraded. Upgraded and downgraded modules will
have their old version and new version displayed, while added and
removed modules display the version of the module that was added or
removed.

!feature

* **Print changes made to Puppetfile when adding modules**
  ([#2230](https://github.com/puppetlabs/bolt/issues/2230))

  The `bolt module add` command and `Add-BoltModule` cmdlet now display
  a message describing changes made to the Puppetfile, including modules
  that have been added, removed, upgraded, or downgraded.