(bug-fix) Ensure inventoryfile is a pathname before passing to outputter
Previously, the inventoryfile passed to the outputter for printing when
running `bolt inventory show` with the Human outputter could be either a
string or a Pathname, depending on whether the user had configured the
inventory location (string) or was using the default project inventory
(Pathname). We erroneously assumed it would always be a
Pathname, which resulted in a 'method not found' error when we tried
calling `exist?` on the parameter. This uses the `File.exist?` method to
verify the path exists which will work for either object passed to the
method.

!bug

* **Fix 'method not found' error running 'bolt inventory show'**

  Previously, when running 'bolt inventory show' with a configured
  inventory path and the human format a 'method not found' error was
  raised. This now correctly prints the targets in the inventory.