(Docs) Harmful terminology changes
I grepped through the docs and found a couple of lingering terms
that we need to change. Most are examples in the docs. One is a DB
filename that is going to change soon. I'm on the fence about merging
this now or waiting for the Puppet 7 release when all of the other docs
changes (and a lot of the engineering work) will land.

!no-release-note