Allow whitespace in heredocs with Rubocop
Rubocop recently updated to 1.0.0, which is flagging whitespace in some
Heredocs in the repo. This configures rubocop to allow whitespace in
heredocs, since that should always be a false positive for our use cases.

This also updates the Gemfile to pull in the latest rubocop in order for
local rubocop runs to match CI.