(GH-2111) Shadow fact/variable collisions in PAL catalog compiler
This updates the `Bolt::Catalog.compile_catalog` method to pass target
and plan variables to the catalog compiler, which will then deserialize
plan variables and shadow any variables that collide with facts of the
same name. This change fixes a bug where a plan variable 'foo' that
collided with a fact 'foo' would not be passed to the compiler, so any
plan variable 'bar' that had a local reference to 'foo' would not be
deserialized correctly.

!bug

* **Correctly shadow fact/variable collisions in apply blocks**
  ([#2111](https://github.com/puppetlabs/bolt/issues/2111))

  Bolt now correctly shadows target and plan variables that collide with
  facts of the same name when running apply blocks.