(GH-2219) Add 'module' and 'modulepath' topic guides
This adds a new `module` topic guide. It gives a brief description of
what a module is and how Bolt can manage a project's modules.

This also adds a new 'modulepath' topic guide that describes what the
modulepath is, precedence of modules on the modulepath, and how Bolt
adds directories to the modulepath.

!no-release-note