(GH-2245) Safely delete tmpdir used when configuring Puppet for PAL
This updates `Bolt::PAL.with_puppet_settings` to safely delete the
tmpdir used to configure Puppet. Previously, the tmpdir was created with
a block, which would result in the tmpdir being automatically deleted
once the block completed. However, if the tmpdir was deleted before this
happened, Bolt would error and show a stacktrace. Now, Bolt manually
deletes the tmpdir if it exists instead of relying on the `mktmpdir`
method to do it automatically.

!bug

* **Safely delete tmpdir used to configure Puppet for PAL**
  ([#2245](https://github.com/puppetlabs/bolt/issues/2245))

  Bolt now safely deletes the tmpdir used to configure Puppet when using
  PAL. Previously, if the tmpdir was deleted during a Bolt run before
  Bolt deleted the directory itself, an error with a stacktrace would be
  raised.