(GH-2268) Add 'wait_until_available' method for puppetserver tests
This adds a `wait_until_available` method to `BoltSpec::BoltServer` that
will wait until puppetserver is running and accepting requests.
This fixes an issue where tests might be run too soon after the
puppetserver container was started and before it was ready to accept
requests, causing the tests to fail.

!no-release-note