(GH-2057) Add file and line number to Bolt errors
This gets the file and line number from the Puppet plan function and
passes it all the way through to the Result object, so that if an error
occurs we can include the file and line number it came from under the
`details` key of the result. This ensures that users can find where a
particular failure is coming from not just from Puppet errors (ie. the
wrong number of arguments to a function) but also to Bolt errors (ie.  a
task that exits 1). This gets the file and line number simply using
`Puppet::Pops::PuppetStack.top_of_stack`. This only modifies the Result
objects for Results that build an error hash: `run_task`, `run_command`,
and `run_script`, as well as failures from `upload_file` and
`download_file`. This does not currently work for YAML plans.

Closes #2057

!feature

* **File and line number included in plan function errors** ([2057](https://github.com/puppetlabs/bolt/issues/2057))

  If the plan functions `run_command`, `run_script`, or `run_task` fail
  they will now include the file and line number in the `details` key of
  the Result object. This information will also be printed when run with
  info level logging or higher.