(GH-2171) Do not override SSL variables in PowerShell module
This updates the PowerShell module to not override the `SSL_CERT_FILE`
and `SSL_CERT_DIR` environment variables if they are already set.
Previously, the PowerShell module would always override these
environment variables, which could lead to SSL errors if a user
attempted to provide a cert on the command line.
are not already set.

!bug

* **Do not override SSL variables in PowerShell module**
  ([#2171](https://github.com/puppetlabs/bolt/issues/2171))

  The PowerShell module no longer overrides the `SSL_CERT_FILE` and
  `SSL_CERT_DIR` environment variables if they are already set.