(GH-2278) Include file and line number in YAML plan eval errors
This adds the file path and line number of an error for evaluated code
in a YAML plan. Previously, if evaluated code resulted in an error, the
error would include no file path and the line number of the error within
the code itself (as though the code were in its own file). This made
these errors less helpful, as they did not correctly indicate where an
error in a YAML plan had occurred.

!feature

* **Include file and line number in YAML plan code evaluation errors**
  ([#2278](https://github.com/puppetlabs/bolt/issues/2278))

  Errors raised when evaluating code in a YAML plan now include the path
  to the YAML plan and the line number that the error occurred on in the
  plan.

!removal

* **Folded scalar values in YAML plans no longer evaluated**
  ([#2306](https://github.com/puppetlabs/bolt/pull/2306))

  Folded scalar values in YAML plans are no longer evaluated and are
  instead treated as string literals.