(GH-2127) Don't thrash while waiting for parallel tasks to finish
This modifies how the executor manages several yarns running in
parallel. Originally we would continuously loop through the yarns and
see if the task had finished by resuming the Fiber, checking the thread,
and yielding if the thread was still executing. This loop ran
continuously while waiting for results and ate up so many CPU cycles
that a 2-second task would take as long as 2 minutes to complete. The
temporary fix for this was to add a `sleep(0)` to the Fiber check so
that there was a context switch and the CPU could resume executing the
task itself before continuing to loop and check other tasks. This isn't
a great long-term solution though, as it doesn't directly address the
issue (?).

Now the executor subscribes to `:node_result` messages while running in
a parallel block, and unsubscribes from them when exiting the block.
`:node_result` messages are sent from the transport class when an action
finishes executing, indicating to the executor (in this context) that a
thread has finished and it can resume round-robining to resume Fibers.
This allows the executor to do nothing while waiting for results, and
only resume Fibers once it knows one of them has completed.